pyproject.toml
[project]
name = "zerocode-tgbot-builder"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
	"telethon",
	"aiosqlite",
	"pyyaml",
	"tonutils",
	"tonsdk"
]
